+++ test/testrunner.py
@@ -15,11 +15,12 @@ from cases.response.string import *
 from cases.response.text import *
 from cases.response.photo import *
 from cases.after import afterTests, testResultOkCalls
+from cases.commands import *
 
 from constants import DEV_BOT_TOKEN
 
 def getBot(rules):
-    bot = bobot.init(DEV_BOT_TOKEN)
+    bot = bobot.init(DEV_BOT_TOKEN, 'TestBot')
     bot.setWebhook(None)
     bot.rule(rules)
     return bot
@@ -53,7 +54,24 @@ class RuleAfterTestCases(unittest.TestCase):
         testResultOkCalls.checkAfter(bot)
         self.assertNotEqual(testResultOkCalls.before, testResultOkCalls.after)
         self.assertEqual(testResultOkCalls.awaits, testResultOkCalls.after)
-        
+
+class RuleCommandTestCases(unittest.TestCase):
+    def testCommandWithSlash(self):
+        bot = getBot(commandWithSlash.rules)
+        self.assertTrue(commandWithSlash.check(bot))
+    
+    def testCommandWithSlashAndUserName(self):
+        bot = getBot(commandWithSlashAndUserName.rules)
+        self.assertTrue(commandWithSlashAndUserName.check(bot))
+    
+    def testCommandStandaloneWithSlash(self):
+        bot = getBot(commandStandaloneWithSlash.rules)
+        self.assertTrue(commandStandaloneWithSlash.check(bot))
+    
+    def testCommandStandaloneWithSlashAndUserName(self):
+        bot = getBot(commandStandaloneWithSlashAndUserName.rules)
+        self.assertTrue(commandStandaloneWithSlashAndUserName.check(bot))
+
 class TextResponseTestCases(unittest.TestCase):
     def testResponseAsText(self):
         bot = getBot(responseAsText.rules)