+++ Changelog.md
@@ -4,6 +4,8 @@
 
 ### 0.1.4
  - + Added `after` rule property
+ - + Added `command` rule property
+ - + Added test for `after` and `command` properties in Rule
 
 ### 0.1.3
  - * Fixed missing Sticker class in exports